frontend: Move the settings into the config redux
Settings are not necessarily related to the UI, but they're related to
configuration, so this patch moves the settings into the config redux
entries of the store.
