+++ frontend/src/components/App/settings/hook.tsx
@@ -3,7 +3,7 @@ import { useTypedSelector } from '../../../redux/reducers/reducers';
 
 export const useSettings = function (settingName?: string) {
   const storeSettingEntries = useTypedSelector(state =>
-    settingName ? state.ui.settings[settingName] : state.ui.settings
+    settingName ? state.config.settings[settingName] : state.config.settings
   );
   const [settingEntries, setSettingEntries] = useState(storeSettingEntries);
