+++ frontend/src/components/common/Resource/ResourceTable.stories.tsx
@@ -22,7 +22,7 @@ const TemplateWithFilter: Story<{
     (
       state = {
         filter: { namespaces: new Set<string>(), search: '' },
-        ui: { settings: { tableRowsPerPageOptions: [10, 20, 50, 100] } },
+        config: { settings: { tableRowsPerPageOptions: [10, 20, 50, 100] } },
       }
     ) => state,
     {
@@ -30,7 +30,7 @@ const TemplateWithFilter: Story<{
         namespaces: new Set(namespaces),
         search,
       },
-      ui: {
+      config: {
         settings: {
           tableRowsPerPageOptions: [10, 20, 50, 100],
         },
