+++ frontend/src/redux/actions/actions.tsx
@@ -33,7 +33,7 @@ export const UI_BRANDING_SET_APP_LOGO = 'UI_BRANDING_SET_APP_LOGO';
 export const UI_SET_CLUSTER_CHOOSER_BUTTON = 'UI_SET_CLUSTER_CHOOSER_BUTTON';
 export const UI_HIDE_APP_BAR = 'UI_HIDE_APP_BAR';
 export const UI_FUNCTIONS_OVERRIDE = 'UI_FUNCTIONS_OVERRIDE';
-export const UI_SET_SETTINGS = 'UI_SET_SETTINGS';
+export const CONFIG_SET_SETTINGS = 'CONFIG_SET_SETTINGS';
 
 export interface BrandingProps {
   logo: AppLogoType;
@@ -201,6 +201,6 @@ export function setFunctionsToOverride(override: FunctionsToOverride) {
   return { type: UI_FUNCTIONS_OVERRIDE, override };
 }
 
-export function setUISettings(settings: { [key: string]: any }) {
-  return { type: UI_SET_SETTINGS, settings };
+export function setAppSettings(settings: { [key: string]: any }) {
+  return { type: CONFIG_SET_SETTINGS, settings };
 }
